Update main components
