+++ src/components/app/app.tsx
@@ -1,17 +1,15 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck баг с react router
 import { Route, Routes } from 'react-router-dom';
 import { AppRoute } from '../../const';
 import { OfferPartial } from '../../lib/types/offer';
 import { City } from '../../lib/types/city';
 import { browserHistory } from '../../browser-history';
-import Main from '../../pages/main/main';
 import Login from '../../pages/login/login';
 import Favorites from '../../pages/favorites/favorites';
 import Offer from '../../pages/offer/offer';
 import NotFound from '../../pages/not-found/not-found';
 import PrivateRoute from '../private-route/private-route';
 import HistoryRouter from '../history-router/history-router';
+import Main from '../../pages/main/main';
 
 type Props = {
   favorites: OfferPartial[];
@@ -29,7 +27,7 @@ const App = ({ favorites, cityList }: Props) => (
       <Route
         path={AppRoute.Favorites}
         element={
-          //@ts-expect-error происходит какое-то безумие
+          // @ts-expect-error происходит какое-то безумие (не проходит сборка)
           <PrivateRoute>
             <Favorites favorites={favorites} />
           </PrivateRoute>
