+++ src/pages/main/main.tsx
@@ -1,49 +1,33 @@
 import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { useAppDispatch } from '../../store';
-import { RootState } from '../../lib/types/store';
 import {
   getSortedOffers,
   selectFilteredOffers,
   selectOffersResponseStatus,
-  selectSelectedOffer,
 } from '../../store/offers/offers.selectors';
-import { getOffers } from '../../store/api-actions';
 import { ResponseStatus } from '../../const';
-import { selectActiveCity } from '../../store/city/city.selectors';
+import { RootState } from '../../lib/types/store';
 import { selectPlacesSorting } from '../../store/sorting/sorting.selectors';
-import { selectCurrentOffer } from '../../store/offers/offersSlice';
-import Map from '../../components/map/map';
-import CardList from '../../components/card-list/card-list';
-import MainLayout from '../../components/main-layout/main-layout';
-import PlacesSorting from '../../components/places-sorting/places-sorting';
+import { getOffers } from '../../store/api-actions';
+import Header from '../../components/header/header';
+import CityList from '../../components/city-list/city-list';
 import Spinner from '../../components/spinner/spinner';
 import ErrorMessage from '../../components/error-message/error-message';
+import MainFull from '../../components/main-full/main-full';
+import MainEmpty from '../../components/main-empty/main-empty';
+import clsx from 'clsx';
 
-type Props = {
-  cityList: string[];
-};
+type Props = { cityList: string[] };
 
 const Main = ({ cityList }: Props) => {
-  const activeCity = useSelector(selectActiveCity);
   const placesSorting = useSelector(selectPlacesSorting);
   const offers = useSelector((state: RootState) =>
     getSortedOffers(selectFilteredOffers(state), placesSorting)
   );
   const offersResponseStatus = useSelector(selectOffersResponseStatus);
-  const selectedOffer = useSelector(selectSelectedOffer);
 
   const dispatch = useAppDispatch();
-
-  const handleOfferHover = (offerId: string | undefined) => {
-    if (offerId === undefined) {
-      dispatch(selectCurrentOffer());
-    }
-
-    const currentOffer = offers.find((offer) => offer.id === offerId);
-
-    dispatch(selectCurrentOffer(currentOffer));
-  };
 
   useEffect(() => {
     dispatch(getOffers());
@@ -51,39 +35,46 @@ const Main = ({ cityList }: Props) => {
 
   return (
     <div className="page page--gray page--main">
-      <MainLayout cityList={cityList}>
-        <section className="cities__places places">
-          <h2 className="visually-hidden">Places</h2>
-          <b className="places__found">
-            {offers.length} places to stay in {activeCity.name}
-          </b>
+      <Header />
 
-          <PlacesSorting />
+      <main
+        className={clsx(
+          'page__main',
+          'page__main--index',
+          !offers.length && 'page__main--index-empty'
+        )}
+      >
+        <h1 className="visually-hidden">Cities</h1>
 
-          {offersResponseStatus === ResponseStatus.Pending && <Spinner />}
-
-          {offersResponseStatus === ResponseStatus.Error && (
-            <ErrorMessage message="Failed to load data" />
-          )}
-
-          {offersResponseStatus === ResponseStatus.Success && (
-            <CardList
-              offers={offers}
-              cardType="offer"
-              onCardMouseOver={handleOfferHover}
-            />
-          )}
-        </section>
-        <div className="cities__right-section">
-          <section className="cities__map map">
-            <Map
-              city={activeCity}
-              locations={offers.map((offer) => offer.location)}
-              selectedLocation={selectedOffer?.location}
-            />
+        <div className="tabs">
+          <section className="locations container">
+            <CityList cityList={cityList} />
           </section>
         </div>
-      </MainLayout>
+
+        <div className="cities">
+          <div
+            className={clsx(
+              'cities__places-container',
+              'container',
+              !offers.length && 'cities__places-container--empty'
+            )}
+          >
+            {offersResponseStatus === ResponseStatus.Pending && <Spinner />}
+
+            {offersResponseStatus === ResponseStatus.Error && (
+              <ErrorMessage message="Failed to load data" />
+            )}
+
+            {offersResponseStatus === ResponseStatus.Success &&
+            offers.length ? (
+              <MainFull />
+            ) : (
+              <MainEmpty />
+            )}
+          </div>
+        </div>
+      </main>
     </div>
   );
 };
